refactor: InputValidator의 역할 InputView에서 진행하도록 변경
