+++ src/main/java/view/InputView.java
@@ -5,9 +5,6 @@ import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
 
-import static view.InputViewValidator.validateLadderHeight;
-import static view.InputViewValidator.validatePlayerNames;
-
 public class InputView {
 
     public static final String INPUT_PLAYER_NAMES_MESSAGE = "참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)";
@@ -18,7 +15,6 @@ public class InputView {
     public static List<String> readPlayerNames() {
         printInputMessage(INPUT_PLAYER_NAMES_MESSAGE);
         String playerNamesInput = readLine();
-        validatePlayerNames(playerNamesInput);
         return Arrays.stream(playerNamesInput.split(",", -1))
                 .map(String::strip)
                 .collect(Collectors.toUnmodifiableList());
@@ -27,7 +23,7 @@ public class InputView {
     public static int readLadderHeight() {
         printInputMessage(INPUT_LADDER_HEIGHT_MESSAGE);
         String ladderHeight = readLine();
-        validateLadderHeight(ladderHeight);
+        validateNotInteger(ladderHeight);
         return Integer.parseInt(ladderHeight);
     }
 
@@ -36,7 +32,23 @@ public class InputView {
     }
 
     private static String readLine() {
-        return scanner.nextLine().strip();
+        String inputValue = scanner.nextLine().strip();
+        validateNotEmpty(inputValue);
+        return inputValue;
+    }
+
+    private static void validateNotEmpty(final String input) {
+        if (input.isBlank()) {
+            throw new IllegalArgumentException("입력값이 비어있습니다.");
+        }
+    }
+
+    public static void validateNotInteger(final String input) {
+        try {
+            Integer.parseInt(input);
+        } catch (NumberFormatException exception) {
+            throw new IllegalArgumentException("사다리 높이는 숫자만 입력할 수 있습니다.");
+        }
     }
 
 }
